DRA: plumb Provider to StaticAutoscaler, save a Snapshot along ClusterSnapshot every loop
clustersnapshot.Handle is introduced to group ClusterSnapshot and the
DRA object Snapshot needed for adding pods/node to it. ClusterSnapshot
is inlined into the Handle, and AutoscalingContext.ClusterSnapshot is
now a Handle. Thanks to this none of the read accesses to the field
need to be updated.

Putting the DRA object snapshot directly into AutoscalingContext along
ClusterSnapshot is not an option, since there are places importing
clustersnapshot that can't import context because context imports them.

ClusterSnapshot is also now initialized using the DRA object snapshot.